Add .5 law-pass layers: distribute law-concepts earlier
Law-concepts (assoc, commut, hasId, hasInv, closes, distrib) were
defined only in L14 but their expanded forms existed from L05. The
engine couldn't connect expanded laws to named concepts across 9 layers.

Five .5 companion layers now define law-concepts at their first instance
and restate expanded laws in compact named form. Cross-domain
universality is structurally visible: assoc spans {+,×,∘,⊻}, commut
spans {+,×,gcd,⊻}, hasId spans 7 operators with 4 distinct identities.

- 05.5: assoc, commut, hasId, distrib defs + compact forms for +,×
- 06.5: closes def + hasId(∪,∅)
- 07.5: assoc(∘), hasId(∘,ℑ)
- 09.5: hasInv def + assoc/commut/hasId/hasInv for ⊻
- 13.5: hasId/commut for gcd, hasId for lcm
- L05: add × assoc and × commut expanded forms
- L07: add ∘ composition identity expanded form
- L14: remove 19 lines moved to .5 layers (36→16)
- build_db.py: float layer numbers, REAL column type

+++ tools/build_db.py
@@ -67,7 +67,7 @@ def load_layers():
     """Load all triples from JSONL layer files."""
     triples = []
     for path in sorted(glob.glob(os.path.join(LAYERS_DIR, "*.jsonl"))):
-        layer = int(os.path.basename(path).split("_")[0])
+        layer = float(os.path.basename(path).split("_")[0])
         with open(path, "r", encoding="utf-8") as f:
             for line_num, line in enumerate(f, 1):
                 line = line.strip()
@@ -197,7 +197,7 @@ def build_database(symbols, triples):
             glyph TEXT PRIMARY KEY,
             name TEXT NOT NULL,
             role TEXT NOT NULL,
-            layer INTEGER NOT NULL,
+            layer REAL NOT NULL,
             depth INTEGER NOT NULL DEFAULT 0,
             token TEXT NOT NULL UNIQUE
         );
@@ -207,7 +207,7 @@ def build_database(symbols, triples):
             subject TEXT NOT NULL,
             operator TEXT NOT NULL,
             object TEXT NOT NULL,
-            layer INTEGER NOT NULL,
+            layer REAL NOT NULL,
             line INTEGER NOT NULL,
             form TEXT NOT NULL,
             depth INTEGER NOT NULL DEFAULT 0
